# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Clubs SaaS API",
        "description": "API documentation for Clubs SaaS",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/Clubs/public/api",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/v1/player-subscriptions": {
            "get": {
                "tags": [
                    "Subscription Management"
                ],
                "summary": "👥 List all player subscriptions",
                "operationId": "078e2cb73b70d3a434d3424d990aec51",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Subscription Management"
                ],
                "summary": "➕ Subscribe a member to a plan",
                "operationId": "eda32873f1191700b574fa617b5080fd",
                "responses": {
                    "201": {
                        "description": "Subscription created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/player-subscriptions/{id}": {
            "get": {
                "tags": [
                    "Subscription Management"
                ],
                "summary": "🔍 Get subscription details",
                "operationId": "8a9b280572c05b196fa3a2805e94eac7",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/player-subscriptions/{id}/freeze": {
            "post": {
                "tags": [
                    "Subscription Management"
                ],
                "summary": "❄️ Freeze a subscription",
                "operationId": "31e46dfe8b6d3d1ec5b1b87e0bf4ee11",
                "responses": {
                    "200": {
                        "description": "Subscription frozen"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/subscription-plans": {
            "get": {
                "tags": [
                    "Subscription Management"
                ],
                "summary": "📋 List all subscription plans",
                "operationId": "d777883f9934526c98bba1e8b8912722",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Subscription Management"
                ],
                "summary": "➕ Create a new plan",
                "operationId": "d30a88fbdf1b0755a4820d63d239814a",
                "responses": {
                    "201": {
                        "description": "Plan created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/subscription-plans/{id}": {
            "get": {
                "tags": [
                    "Subscription Management"
                ],
                "summary": "🔍 Get plan details",
                "operationId": "e24d7c05878abe2afc0ac3932f9d872a",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Subscription Management"
                ],
                "summary": "📝 Update plan",
                "operationId": "1f405e846e1f12740ba00fd943caf15a",
                "responses": {
                    "200": {
                        "description": "Plan updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Subscription Management"
                ],
                "summary": "🗑️ Delete plan",
                "operationId": "4d39b9cd9a51b2220ec1907ebdf76e6f",
                "responses": {
                    "200": {
                        "description": "Plan deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}